remove domains module, move all database download logic to download module
