+++ synthaser/parsers.py
@@ -10,16 +10,10 @@ from synthaser import __version__
 def add_rules_group(parser):
     group = parser.add_argument_group("Other arguments")
     group.add_argument(
-        "-df",
-        "--domain_file",
+        "-uf",
+        "--rule_file",
         type=argparse.FileType(),
-        help="JSON file containing alternate domain targets",
-    )
-    group.add_argument(
-        "-cf",
-        "--classify_file",
-        type=argparse.FileType(),
-        help="JSON file containing custom classification rules"
+        help="Custom rule file generated by the synthaser rule generator",
     )
 
 
@@ -200,14 +194,17 @@ def add_getdb_subparser(subparsers):
     p = subparsers.add_parser(
         "getdb",
         help="Download a CDD database for local searches",
-        description="Download a pre-formatted rpsblast database."
-        " For full description of the available databases, see:"
-        " https://www.ncbi.nlm.nih.gov/Structure/cdd/cdd_help.shtml#CDSource",
+        description="Download a pre-formatted rpsblast database.\n\n"
+        "For full description of the available databases, see:\n"
+        " https://www.ncbi.nlm.nih.gov/Structure/cdd/cdd_help.shtml#CDSource\n\n"
+        "Note that 'cdd_families' will download a file containing a summary of\n"
+        "all families in the CDD for rule building - not a searchable database.",
         epilog="Cameron L.M. Gilchrist 2019",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
     )
     p.add_argument(
         "database",
-        choices=["Cdd", "Cdd_NCBI", "Cog", "Kog", "Pfam", "Prk", "Smart", "Tigr"],
+        choices=["cdd_families", "Cdd", "Cdd_NCBI", "Cog", "Kog", "Pfam", "Prk", "Smart", "Tigr"],
         help="Database to be downloaded",
     )
     p.add_argument(
@@ -215,130 +212,6 @@ def add_getdb_subparser(subparsers):
         help="Folder where database is to be saved. Will save a .tar.gz file, and"
         " extract its contents to a folder of the same name.",
     )
-
-
-def add_domain_file_argument(parser):
-    parser.add_argument(
-        "domain_file",
-        type=argparse.FileType("r"),
-        help="CDD family file generated by rulegen parse",
-    )
-
-
-def add_domain_rule_argument(parser):
-    parser.add_argument(
-        "rule_file",
-        help="synthaser Domain rule file"
-    )
-
-
-def add_convert_subparser(subparsers):
-    text = "Convert CDD family names to accessions and vice versa."
-    parser = subparsers.add_parser(
-        "convert",
-        help=text,
-        description=text,
-        epilog="Cameron L.M. Gilchrist, 2020."
-    )
-    add_domain_file_argument(parser)
-    inputs = parser.add_mutually_exclusive_group()
-    inputs.add_argument(
-        "-n",
-        "--names",
-        nargs="+",
-        help="Domain family names"
-    )
-    inputs.add_argument(
-        "-a",
-        "--accessions",
-        nargs="+",
-        help="Domain family accessions"
-    )
-
-
-def add_update_domain_subparser(subparsers):
-    text = "Create domain rule files, add new rules or update existing rules."
-    parser = subparsers.add_parser("update", help=text, description=text)
-    add_domain_file_argument(parser)
-    add_domain_rule_argument(parser)
-
-    single = parser.add_argument_group("Add single rule")
-    single.add_argument("--type", help="Domain type")
-    single.add_argument("--name", help="Domain name")
-    single.add_argument("--families", help="CDD family names", nargs="+")
-
-    multi = parser.add_argument_group("Add multiple rules")
-    multi.add_argument("--file", type=argparse.FileType("r"), help="Skeleton file")
-
-
-def add_remove_rule_subparser(subparsers):
-    parser = subparsers.add_parser(
-        "remove",
-        help="Remove domain rules, or specific CDD families from rules.",
-        description="Remove domain rules, or specific CDD families from rules."
-        " Either remove multiple entire rules, or specify a single rule and multiple"
-        " CDD family accessions using --families."
-    )
-    add_domain_rule_argument(parser)
-    parser.add_argument("--rules", default=[], nargs="+", help="Rule classes")
-    parser.add_argument("--families", default=[], nargs="+", help="CDD family accessions")
-
-
-def add_download_cdd_subparser(subparsers):
-    text = "Download CDD family information from the NCBI."
-    parser = subparsers.add_parser(
-        "download",
-        help=text,
-        description=text,
-    )
-    parser.add_argument("domain_file", help="Output CDD family file")
-    parser.add_argument("--folder", help="Output folder")
-    parser.add_argument("--indent", default=2, help="JSON indent level (def. 2)")
-
-
-def add_summary_subparser(subparsers):
-    text = "Print out domain rules."
-    parser = subparsers.add_parser(
-        "summary",
-        help=text,
-        description=text + " If no rules or families are specified, this module"
-        " will print a summary of the current rules.",
-    )
-    add_domain_rule_argument(parser)
-    parser.add_argument("--rules", default=[], nargs="+", help="Rule classes")
-    parser.add_argument("--families", default=[], nargs="+", help="CDD family accessions")
-
-
-def add_domain_rules_subparser(subparsers):
-    text = "Create or manipulate domain rule files"
-    parser = subparsers.add_parser(
-        "domains",
-        help=text,
-        description=text,
-        epilog="Usage examples\n--------------\n"
-        "Download and build CDD family database:\n"
-        "  $ synthaser domains download cdd.json\n\n"
-        "Create a new domain rule file from a skeleton:\n"
-        "  $ synthaser domains update cdd.json myrules.json \\\n"
-        "      --file skeleton.txt\n\n"
-        "Add the PKS_KS domain family as a KS rule:\n"
-        "  $ synthaser domains update cdd.json myrules.json \\\n"
-        "      --type KS --name PKS_KS\n\n"
-        "...and remove it:\n"
-        "  $ synthaser domains remove myrules.json --families PKS_KS\n\n"
-        "Get the CDD accession for PKS_KS (smart00825):\n"
-        "  $ synthaser domains convert cdd.json --names PKS_KS\n\n"
-        "Print a summary of a domain rule file:\n"
-        "  $ synthaser domains summary myrules.json\n\n"
-        "Cameron L.M. Gilchrist, 2020.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    subparsers = parser.add_subparsers(dest="subcommand")
-    add_download_cdd_subparser(subparsers)
-    add_update_domain_subparser(subparsers)
-    add_remove_rule_subparser(subparsers)
-    add_convert_subparser(subparsers)
-    add_summary_subparser(subparsers)
 
 
 def add_extract_subparser(subparsers):
@@ -406,7 +279,6 @@ def get_parser():
     )
     subparsers = parser.add_subparsers(dest="command")
     add_search_subparser(subparsers)
-    add_domain_rules_subparser(subparsers)
     add_getdb_subparser(subparsers)
     add_getseq_subparser(subparsers)
     add_extract_subparser(subparsers)
